@@ -128,7 +128,7 @@ class FlowyButton extends StatelessWidget {
                 (Platform.isIOS || Platform.isAndroid)
             ? BoxDecoration(
                 border: Border.all(
-                color: Theme.of(context).colorScheme.surfaceVariant,
+                color: Theme.of(context).colorScheme.surfaceContainerHighest,
                 width: 1.0,
               ))
             : null);
@@ -210,12 +210,12 @@ class FlowyTextButton extends StatelessWidget {
         onPressed: onPressed ?? () {},
         focusNode: FocusNode(skipTraversal: onPressed == null),
         style: ButtonStyle(
-          overlayColor: const MaterialStatePropertyAll(Colors.transparent),
+          overlayColor: const WidgetStatePropertyAll(Colors.transparent),
           splashFactory: NoSplash.splashFactory,
           tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-          padding: MaterialStateProperty.all(padding),
-          elevation: MaterialStateProperty.all(0),
-          shape: MaterialStateProperty.all(
+          padding: WidgetStateProperty.all(padding),
+          elevation: WidgetStateProperty.all(0),
+          shape: WidgetStateProperty.all(
             RoundedRectangleBorder(
               side: BorderSide(
                 color: isDangerous
@@ -225,7 +225,7 @@ class FlowyTextButton extends StatelessWidget {
               borderRadius: radius ?? Corners.s6Border,
             ),
           ),
-          textStyle: MaterialStateProperty.all(
+          textStyle: WidgetStateProperty.all(
             TextStyle(
               fontWeight: fontWeight ?? FontWeight.w500,
               fontSize: fontSize,
@@ -233,9 +233,9 @@ class FlowyTextButton extends StatelessWidget {
               fontFamily: fontFamily,
             ),
           ),
-          backgroundColor: MaterialStateProperty.resolveWith(
+          backgroundColor: WidgetStateProperty.resolveWith(
             (states) {
-              if (states.contains(MaterialState.hovered)) {
+              if (states.contains(WidgetState.hovered)) {
                 return hoverColor ??
                     (isDangerous
                         ? Theme.of(context).colorScheme.error
@@ -248,9 +248,9 @@ class FlowyTextButton extends StatelessWidget {
                       : Theme.of(context).colorScheme.secondaryContainer);
             },
           ),
-          foregroundColor: MaterialStateProperty.resolveWith(
+          foregroundColor: WidgetStateProperty.resolveWith(
             (states) {
-              if (states.contains(MaterialState.hovered)) {
+              if (states.contains(WidgetState.hovered)) {
                 return fontHoverColor ??
                     (fontColor ?? Theme.of(context).colorScheme.onSurface);
               }
