@@ -102,7 +102,7 @@ class _EventCardState extends State<EventCard> {
           hoverColor: Theme.of(context).brightness == Brightness.light
               ? const Color(0x0F1F2329)
               : const Color(0x0FEFF4FB),
-          foregroundColorOnHover: Theme.of(context).colorScheme.onBackground,
+          foregroundColorOnHover: Theme.of(context).colorScheme.onSurface,
         ),
       ),
       onStartEditing: () {},
