@@ -300,7 +300,7 @@ class ColorOptionAction extends PopoverActionCell {
           colors: colors,
           selected: selectedColor,
           border: Border.all(
-            color: Theme.of(context).colorScheme.onBackground,
+            color: Theme.of(context).colorScheme.onSurface,
           ),
           onTap: (option, index) async {
             final transaction = editorState.transaction;
