@@ -77,8 +77,7 @@ class _PageStyleIconState extends State<PageStyleIcon> {
       showDoneButton: true,
       showHeader: true,
       title: LocaleKeys.titleBar_pageIcon.tr(),
-      backgroundColor: Theme.of(context).colorScheme.background,
-      isScrollControlled: true,
+      backgroundColor: Theme.of(context).colorScheme.surface,
       enableDraggableScrollable: true,
       minChildSize: 0.6,
       initialChildSize: 0.61,
