@@ -208,8 +208,7 @@ class _FontButton extends StatelessWidget {
       showDoneButton: true,
       showHeader: true,
       title: LocaleKeys.titleBar_font.tr(),
-      backgroundColor: Theme.of(context).colorScheme.background,
-      isScrollControlled: true,
+      backgroundColor: Theme.of(context).colorScheme.surface,
       enableDraggableScrollable: true,
       minChildSize: 0.6,
       initialChildSize: 0.61,
